+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -11,6 +11,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
+@Transactional(readOnly = true)
 public class TableGroupService {
 
     private final TableGroupRepository tableGroupRepository;
