+++ src/main/resources/application-test.properties
@@ -6,5 +6,9 @@ spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.show-sql=true
 
 spring.jpa.hibernate.ddl-auto=create-drop
-logging.level.org.hibernate.type.descriptor.sql = trace
+logging.level.org.hibernate.type.descriptor.sql=trace
+
+
+spring.jpa.open-in-view=false
+
