+++ src/test/java/kitchenpos/ordertable/acceptance/step/TableAcceptanceStep.java
@@ -7,7 +7,7 @@ import io.restassured.mapper.TypeRef;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
-import kitchenpos.acceptance.step.HttpUtil;
+import kitchenpos.util.HttpUtil;
 import kitchenpos.ordertable.dto.OrderTableRequest;
 import kitchenpos.ordertable.dto.OrderTableResponse;
