+++ src/test/java/kitchenpos/product/acceptance/step/ProductAcceptanceStep.java
@@ -8,7 +8,7 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.acceptance.step.HttpUtil;
+import kitchenpos.util.HttpUtil;
 import kitchenpos.product.dto.ProductRequest;
 import kitchenpos.product.dto.ProductResponse;
