+++ src/test/java/kitchenpos/util/HttpUtil.java
@@ -1,4 +1,4 @@
-package kitchenpos.acceptance.step;
+package kitchenpos.util;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
